+++ src/console/src/guards.rs
@@ -1,7 +1,7 @@
 use crate::STATE;
 use ic_cdk::caller;
 use shared::controllers::is_controller;
-use shared::types::interface::Controllers;
+use shared::types::state::Controllers;
 
 pub fn caller_is_controller() -> Result<(), String> {
     let caller = caller();
