+++ src/mission_control/src/lib.rs
@@ -23,7 +23,8 @@ use ic_cdk::{storage, trap};
 use ic_cdk_macros::{init, post_upgrade, pre_upgrade, query, update};
 use ic_ledger_types::Tokens;
 use satellites::store::get_satellites;
-use shared::types::interface::{Controllers, MissionControlArgs};
+use shared::types::interface::MissionControlArgs;
+use shared::types::state::Controllers;
 use shared::types::state::UserId;
 use std::cell::RefCell;
 use std::collections::{HashMap, HashSet};
