+++ src/mission_control/src/types.rs
@@ -1,6 +1,6 @@
 pub mod state {
     use candid::{CandidType, Deserialize, Principal};
-    use shared::types::interface::Controllers;
+    use shared::types::state::Controllers;
     use shared::types::state::UserId;
     use std::collections::HashMap;
