+++ src/satellite/src/lib.rs
@@ -55,7 +55,8 @@ use ic_cdk::storage::{stable_restore, stable_save};
 use ic_cdk_macros::{init, post_upgrade, pre_upgrade, query, update};
 use rules::constants::DEFAULT_DB_COLLECTIONS;
 use shared::constants::MAX_NUMBER_OF_SATELLITE_CONTROLLERS;
-use shared::types::interface::{Controllers, ControllersArgs, SatelliteArgs};
+use shared::types::interface::{ControllersArgs, SatelliteArgs};
+use shared::types::state::Controllers;
 use std::cell::RefCell;
 use std::collections::{BTreeMap, HashMap, HashSet};
 use types::list::ListParams;
