+++ src/satellite/src/upgrade/types.rs
@@ -8,7 +8,7 @@ pub mod upgrade {
     use crate::storage::types::state::StorageStableState;
     use candid::CandidType;
     use serde::Deserialize;
-    use shared::types::interface::Controllers;
+    use shared::types::state::Controllers;
 
     #[derive(Default, CandidType, Deserialize, Clone)]
     pub struct UpgradeStableState {
