+++ src/shared/src/controllers.rs
@@ -1,5 +1,4 @@
-use crate::types::interface::Controllers;
-use crate::types::state::UserId;
+use crate::types::state::{Controllers, UserId};
 use crate::utils::principal_equal;
 use candid::Principal;
